Выгрузка CSV в базу

+++ GraphBackend.Infrastructure/JwtTokenGenerator.cs
@@ -20,7 +20,7 @@ public class JwtTokenGenerator(
         {
             new Claim(JwtRegisteredClaimNames.Sub, userId.ToString()),
             new Claim(JwtRegisteredClaimNames.Email, email),
-            new Claim(ClaimTypes.Role, ((int)role).ToString())
+            new Claim(ClaimTypes.Role, role.ToString())
         };
 
         var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_settings.Secret));
